feat(llm): implement provider-agnostic AI integration layer
- Defined AIClient interface to abstract AI provider functionality
- Refactored OpenAIClient to conform to the AIClient interface
- Introduced factory pattern (NewAIClient) for dynamic provider selection
- Updated config to support pluggable provider selection
- Enables easy extension for future AI providers (e.g., Anthropic, Gemini)
