+++ server/.dev.env
@@ -6,6 +6,7 @@ GITHUB_APP_ID=
 
 PORT=8080
 APP_ENV=development
-# OPENAI_API_KEY=your-openai-key
+AI_PROVIDER=openai
+OPENAI_API_KEY=your-openai-key
 # DATABASE_URL=postgres://user:pass@localhost:5432/dbname
 # QUEUE_SIZE=100
